TTOL-47 refactor: date 및 time 관련 유틸 함수 파일 분리
- 기존에 index.ts에 있던 formatTime 함수 또한 dateTimeUtil 파일로 이동
- DayPickerModal에서 사용되는 parseTimeString, applyTimeToDate 함수 생성
- index.ts 삭제
