--- src/app/_utils/index.ts
@@ -1,9 +1,0 @@
-const formatTime = (date: string | Date) => {
-  return new Intl.DateTimeFormat('en-US', {
-    hour: 'numeric',
-    minute: '2-digit',
-    hour12: true,
-  }).format(new Date(date));
-};
-
-export { formatTime };
